+++ dev/match_result_cmdm.R
@@ -6,27 +6,12 @@ Z <- matrix(rnorm(10 * 3), 10, 3)
 
 #
 mdd(X, Y, compute = "C", center = "U")
-mdd(X, Y, compute = "R", center = "U")
-
 mdd(X, Y, compute = "C", center = "D")
-mdd(X, Y, compute = "R", center = "D")
-
-mdc(X, Y, center = "U")
-mdc(X, Y, center = "D")
-
-pmdd(X, Y, Z)
-pmdc(X, Y, Z)
 
 # cd cmi/simulation/ff/ff5_year
-set.seed(12345)
-X <- matrix(rnorm(10 * 3), 10, 3)
-Y <- rnorm(10)
-Z <- matrix(rnorm(10 * 3), 10, 3)
-
-#
 dyn.load("MDD.so")
 
-MDD <- function(N, P, U, V, Q) {
+MDD_U <- function(N, P, U, V, Q) {
   if (!(is.numeric(N) && is.numeric(P) && is.numeric(U) && is.numeric(V) &&
         is.numeric(Q)))
     stop("arguments must be numeric")
@@ -39,9 +24,7 @@ MDD <- function(N, P, U, V, Q) {
   return(list(Q = out$Q))
 }
 
-MDD(10, 3, X, Y, 0)$Q
-
-MDD <- function(N, P, U, V, M, Q) {
+MDD_D <- function(N, P, U, V, M, Q) {
   if (!(is.numeric(N) && is.numeric(P) && is.numeric(U) && is.numeric(V) &&
         is.numeric(M) && is.numeric(Q)))
     stop("arguments must be numeric")
@@ -55,15 +38,39 @@ MDD <- function(N, P, U, V, M, Q) {
   return(list(Q = out$Q))
 }
 
-MDD(10, 3, X, Y, matrix(0, 10, 10), 0)$Q
+set.seed(12345)
+X <- matrix(rnorm(10 * 3), 10, 3)
+Y <- rnorm(10)
+Z <- matrix(rnorm(10 * 3), 10, 3)
 
 #
-source('MDD.R')
+MDD_U(10, 3, X, Y, 0)$Q
+MDD_D(10, 3, X, Y, matrix(0, 10, 10), 0)$Q
 
+#
+set.seed(12345)
+X <- matrix(rnorm(10 * 3), 10, 3)
+Y <- matrix(rnorm(10 * 3), 10, 3)
+Z <- matrix(rnorm(10 * 3), 10, 3)
+
+#
+mdd(X, Y, compute = "R", center = "U")
+mdc(X, Y, center = "U")
+pmdd(X, Y, Z)
+pmdc(X, Y, Z)
+mddm(X, Y)
+
+# cd cmi/simulation/reference
+source('MDD/MDD.R')
+source('MDDM/MDDM.R')
+set.seed(12345)
+X <- matrix(rnorm(10 * 3), 10, 3)
+Y <- matrix(rnorm(10 * 3), 10, 3)
+Z <- matrix(rnorm(10 * 3), 10, 3)
+
+#
 MDD(X, Y)
-
 MDC(X, Y)
-
 pMDD(X, Y, Z)
 pMDC(X, Y, Z)
-
+MDDM(X, Y)